feat: web and api runs from same endpoint
