+++ package.json
@@ -94,6 +94,7 @@
     "jsonwebtoken": "^8.5.1",
     "jwt-decode": "^2.2.0",
     "morgan": "^1.10.0",
+    "next": "9.5.2",
     "next-cookies": "^2.0.3",
     "npm-run-all": "^4.1.5",
     "pluralize": "^8.0.0",
