+++ services/app/public/favicon.ico
